Report filtered parameters even for non-form requests via Action Dispatch params; Creates class for handling Entry logic/data. re: #45
